# pkg/task/controller/taskControllerImpl.go
package controller

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/jakkaphatminthana/go-gin/pkg/custom"
	taskModel "github.com/jakkaphatminthana/go-gin/pkg/task/model"
	_taskService "github.com/jakkaphatminthana/go-gin/pkg/task/service"
)

type taskControllerImpl struct {
	taskService _taskService.TaskService
}

func NewTaskControllerImpl(
	taskService _taskService.TaskService,
) TaskController {
	return &taskControllerImpl{
		taskService: taskService,
	}
}

// Listing
//
//	@Router		/task [get]
//	@Summary	List Tasks
//	@Tags		Tasks
//	@Accept		json
//	@Produce	json
//	@Success	200	{object}	utils.DocsDataObject[taskModel.Task]
//	@Failure	400	{object}	utils.DocsErrorBadRequest
//	@Failure	404	{object}	utils.DocsErrorNotFound
//	@Failure	500	{object}	utils.DocsErrorInternalServerError
func (c *taskControllerImpl) Listing(pctx *gin.Context) {
	taskModelList, err := c.taskService.Listing()
	if err != nil {
		custom.Error(pctx, err, http.StatusInternalServerError)
		return
	}

	custom.Success(pctx, http.StatusOK, taskModelList)
}

// Find by ID
//
//	@Router		/task/{id} [get]
//	@Summary	Find task by ID
//	@Tags		Tasks
//	@Accept		json
//	@Produce	json
//	@Param		id	path		number	true	"id of task"
//	@Success	200	{object}	utils.DocsDataObject
//	@Failure	400	{object}	utils.DocsErrorBadRequest
//	@Failure	404	{object}	utils.DocsErrorNotFound
//	@Failure	500	{object}	utils.DocsErrorInternalServerError

func (c *taskControllerImpl) FindById(pctx *gin.Context) {
	var param = struct {
		ID uint64 `uri:"id" binding:"required"`
	}{}

	// validate and convert
	if err := custom.NewCustomRequest(pctx).BindUri(&param); err != nil {
		custom.Error(pctx, err, http.StatusBadRequest)
		return
	}

	taskModel, err := c.taskService.FindById(param.ID)
	if err != nil {
		custom.Error(pctx, err)
		return
	}

	custom.Success(pctx, http.StatusOK, taskModel)
}

// Create
//
//	@Router		/task [post]
//	@Summary	Create task
//	@Tags		Tasks
//	@Accept		json
//	@Produce	json
//	@Param		request	body		model.TaskCreateReq	true	"Request of Creating Task"
//	@Success	200		{object}	utils.DocsDataObject
//	@Failure	400		{object}	utils.DocsErrorBadRequest
//	@Failure	404		{object}	utils.DocsErrorNotFound
//	@Failure	500		{object}	utils.DocsErrorInternalServerError
func (c *taskControllerImpl) Create(pctx *gin.Context) {
	createReq := new(taskModel.TaskCreateReq)

	// validate
	customRequest := custom.NewCustomRequest(pctx)
	if err := customRequest.BindBody(createReq); err != nil {
		custom.Error(pctx, err, http.StatusBadRequest)
		return
	}

	// creating
	task, err := c.taskService.Create(createReq)
	if err != nil {
		custom.Error(pctx, err)
		return
	}

	custom.Success(pctx, http.StatusCreated, task)
}

// Update
//
//	@Router		/task/{id} [patch]
//	@Summary	Update task
//	@Tags		Tasks
//	@Accept		json
//	@Produce	json
//	@Param		id		path		number				true	"id of task"
//	@Param		request	body		model.TaskUpdateReq	true	"Request of Updating Task"
//	@Success	200		{object}	utils.DocsDataObject
//	@Failure	400		{object}	utils.DocsErrorBadRequest
//	@Failure	404		{object}	utils.DocsErrorNotFound
//	@Failure	500		{object}	utils.DocsErrorInternalServerError
func (c *taskControllerImpl) Update(pctx *gin.Context) {
	var param = struct {
		ID uint64 `uri:"id" binding:"required"`
	}{}

	// validate and convert (param)
	if err := custom.NewCustomRequest(pctx).BindUri(&param); err != nil {
		custom.Error(pctx, err, http.StatusBadRequest)
		return
	}

	taskUpdateReq := new(taskModel.TaskUpdateReq)

	// validate and convert (body)
	customRequest := custom.NewCustomRequest(pctx)
	if err := customRequest.BindBody(taskUpdateReq); err != nil {
		custom.Error(pctx, err, http.StatusBadRequest)
		return
	}

	//updating
	task, err := c.taskService.Update(param.ID, taskUpdateReq)
	if err != nil {
		custom.Error(pctx, err)
		return
	}

	custom.Success(pctx, http.StatusCreated, task)
}

// Delete
//
//	@Router		/task/{id} [delete]
//	@Summary	Delete task
//	@Tags		Tasks
//	@Accept		json
//	@Produce	json
//	@Param		id	path		number	true	"id of task"
//	@Success	200	{object}	utils.DocsSuccessMessage
//	@Failure	400	{object}	utils.DocsErrorBadRequest
//	@Failure	404	{object}	utils.DocsErrorNotFound
//	@Failure	500	{object}	utils.DocsErrorInternalServerError
func (c *taskControllerImpl) Delete(pctx *gin.Context) {
	var param = struct {
		ID uint64 `uri:"id" binding:"required"`
	}{}

	// validate and convert (param)
	if err := custom.NewCustomRequest(pctx).BindUri(&param); err != nil {
		custom.Error(pctx, err, http.StatusBadRequest)
		return
	}

	// deleting
	if err := c.taskService.Delete(param.ID); err != nil {
		custom.Error(pctx, err)
		return
	}

	custom.MetaSuccess(pctx, http.StatusOK, "Delete task successful", nil)
}